move to dedicated initializer
